requirements.txt
# pip install -r requirements.txt

openai>=1.0.0
mysql-connector-python>=8.0.0
